set __text_signature__ from user-defined types from an appropriate docstring (issue 5092)
